Add files via upload
The vendingMachine.java file simulates a vending machine. This class is tested by the vendingMachineTestCoin.java and vendingMachineTestGetChoc.java files using the Junit framework. 
This was a class assignment for Software Project Management, the vendingMachine.java file contains the code to be tested and it was provided by the professor. I created vendingMachineTestCoin.java and VendingMachineTestGetChoc.java files.